requirements.txt
aiohttp>=3.9
fastapi>=0.110
uvicorn>=0.27
websockets>=12.0
brotli>=1.1
qrcode>=7.4
requests>=2.31
Pillow>=10.0
cryptography>=42.0
python-dotenv>=1.0
